-Wall warning removal

+++ modules/boost/simd/reduction/include/boost/simd/toolbox/reduction/functions/simd/common/none.hpp
@@ -26,7 +26,7 @@ namespace boost { namespace simd { namespace ext
 
     BOOST_SIMD_FUNCTOR_CALL(1)
     {
-      for(int i=0; i < boost::simd::meta::cardinal_of<A0>::value; i++)
+      for(size_t i=0; i < boost::simd::meta::cardinal_of<A0>::value; i++)
       {
         if(a0[i]) return 0;
       }
